First steps towards cleaned up code
