+++ scripts/helper_scripts/unipept-database-rs/Cargo.toml
@@ -21,3 +21,5 @@ smartstring = { version = "1.0" }
 strum = "0.25.0"
 strum_macros = "0.25.3"
 uniprot = "0.7.0"
+lazy_static = "1.4.0"
+num_cpus = "1.16.0"
