+++ scripts/helper_scripts/unipept-database-rs/src/dat_parser/utils.rs
@@ -1,0 +1,16 @@
+pub fn write_header() {
+    let fields: [&str; 9] = [
+        "Entry",
+        "Sequence",
+        "Protein names",
+        "Version (entry)",
+        "EC number",
+        "Gene ontology IDs",
+        "Cross-reference (InterPro)",
+        "Status",
+        "Organism ID",
+    ];
+
+    let result_string = fields.join("\t");
+    println!("{}", result_string);
+}
