+++ scripts/helper_scripts/unipept-database-rs/src/lib.rs
@@ -2,3 +2,5 @@ pub mod calculate_lcas;
 pub mod taxons_lineages;
 pub mod taxons_uniprots_tables;
 pub mod utils;
+pub mod dat_parser;
+pub mod uniprot;
