+++ scripts/helper_scripts/unipept-database-rs/src/uniprot/mod.rs
@@ -1,0 +1,15 @@
+/// Enum for the different kinds of databases
+#[derive(clap::ValueEnum, Clone, Debug)]
+pub enum UniprotType {
+    Swissprot,
+    Trembl,
+}
+
+impl UniprotType {
+    pub fn to_str(&self) -> &str {
+        match self {
+            UniprotType::Swissprot => "swissprot",
+            UniprotType::Trembl => "trembl",
+        }
+    }
+}
